Split config parsing into multiple files
