+++ problemtools/config_parser/__init__.py
@@ -3,262 +3,10 @@ import re
 from typing import Any, Callable, Generator, Literal, Pattern, Match, ParamSpec, TypeVar
 from collections import defaultdict, deque
 from copy import deepcopy
-
-
-class SpecificationError(Exception):
-    pass
-
-
-type_mapping = {
-    "string": str,
-    "object": dict,
-    "list": list,
-    "bool": bool,
-    "int": int,
-    "float": float,
-}
-
-type_field_mapping = {
-    "*": ["default", "type", "flags", "parsing"],
-    "string": ["alternatives"],
-    "bool": ["alternatives"],
-    "int": ["alternatives"],
-    "float": ["alternatives"],
-    "object": ["required", "properties"],
-    "list": ["content"],
-}
-
-
-def is_copyfrom(val: Any) -> bool:
-    return isinstance(val, tuple) and val[0] == "copy-from"
-
-
-class Path:
-    INDEXING_REGEX = re.compile(r"^([A-Za-z_0-9\-]+)\[(\d+)\]$")
-
-    @staticmethod
-    def parse(path: str) -> Path:
-        parts = path.split("/")
-        res = []
-        for part in parts:
-            m = Path.INDEXING_REGEX.match(part)
-            if m:
-                res.append(m.group(1))
-                res.append(int(m.group(2)))
-            else:
-                res.append(part)
-        return Path(*res)
-
-    @staticmethod
-    def combine(*parts: str | int | Path) -> Path:
-        res = []
-        for part in parts:
-            if isinstance(part, int):
-                res.append(part)
-                continue
-            if isinstance(part, str):
-                part = Path.parse(part)
-            res.extend(list(part.path))  # type: ignore
-        return Path(*res)
-
-    def __init__(self, *path: str | int) -> None:
-        self.path = path
-
-    def index(self, data: dict) -> Any | None:
-        rv = data
-        for part in self.path:
-            if isinstance(part, int):
-                if not isinstance(rv, list):
-                    return None
-                try:
-                    rv = rv[part]
-                except IndexError:
-                    return None
-            else:
-                if part not in rv:
-                    return None
-                rv = rv[part]
-        return rv
-
-    def spec_path(self) -> Path:
-        res = []
-        for part in self.path:
-            if isinstance(part, str):
-                res.append("properties")
-                res.append(part)
-            elif isinstance(part, int):
-                res.append("content")
-        return Path(*res)
-
-    def data_paths(self, data: dict) -> list[Path]:
-        """Finds all data paths that a spec_path is pointing towards (meaning it will explore all items in lists)"""
-
-        def path_is_not_copyfrom(path: Path) -> bool:
-            return not is_copyfrom(path.index(data))
-
-        out = [Path()]
-        state = "base"
-        for part in self.path:
-            if state == "base":
-                if part == "properties":
-                    state = "object-properties"
-                elif part == "content":
-                    state = "base"
-                    new_out = []
-                    for path in out:
-                        val = path.index(data) or []
-                        if is_copyfrom(val):  # skip copied
-                            continue
-                        assert isinstance(val, list)
-                        new_out.extend(Path.combine(path, i) for i in range(len(val)))
-                    out = new_out
-                    if len(out) == 0:
-                        return []
-                else:
-                    assert False
-            elif state == "object-properties":
-                combined_paths = [Path.combine(path, part) for path in out]
-                out = [*filter(path_is_not_copyfrom, combined_paths)]
-                state = 'base'
-
-        return out
-
-    def up(self, levels=1) -> Path:
-        assert levels > 0
-        return Path(*self.path[:-levels])
-
-    def last_name(self) -> int | str:
-        return self.path[-1]
-
-    def __str__(self) -> str:
-        strings = []
-        for part in self.path:
-            if isinstance(part, int):
-                strings[-1] += f"[{part}]"
-            else:
-                strings.append(part)
-        return "/".join(strings)
-
-    def __repr__(self):
-        return f"Path({self})"
-
-    def __eq__(self, value):
-        return self.path == value.path
-
-    def __hash__(self):
-        return hash(self.path)
-
-
-class AlternativeMatch:
-    def __init__(self, matchstr):
-        raise NotImplementedError("Specialize in subclass")
-
-    def check(self, val) -> bool:
-        raise NotImplementedError("Specialize in subclass")
-
-    @staticmethod
-    def get_matcher(type, matchstr) -> "AlternativeMatch":
-        matchers = {
-            "string": StringMatch,
-            "int": IntMatch,
-            "float": FloatMatch,
-            "bool": BoolMatch,
-        }
-        assert type in matchers
-        return matchers[type](matchstr)
-
-
-class StringMatch(AlternativeMatch):
-    def __init__(self, matchstr):
-        self.regex = re.compile(matchstr)
-
-    def check(self, val) -> bool:
-        return self.regex.match(val)
-
-    def __str__(self) -> str:
-        self.regex.pattern
-
-
-class IntMatch(AlternativeMatch):
-    def __init__(self, matchstr: str | int):
-        if isinstance(matchstr, int):
-            self.start = self.end = matchstr
-            return
-        try:
-            if matchstr.count(":") > 1:
-                raise ValueError
-            if ":" in matchstr:
-                self.start, self.end = [
-                    int(p) if p else None for p in map(str.strip, matchstr.split())
-                ]
-            else:
-                matchstr = matchstr.strip()
-                if not matchstr:
-                    raise SpecificationError("Match string for integer was left empty")
-                self.start = self.end = int(matchstr)
-        except ValueError:
-            raise SpecificationError(
-                'Int match string should be of the form "A:B" where A and B can be parsed as ints or left empty, or a single integer'
-            )
-
-    def check(self, val) -> bool:
-        if not isinstance(val, int):
-            return False
-        if self.start is not None:
-            if val < self.start:
-                return False
-        if self.end is not None:
-            if val > self.start:
-                return False
-        return True
-
-    def __str__(self):
-        if A == B:
-            return str(A)
-        A = str(self.start) if self.start is not None else ""
-        B = str(self.end) if self.end is not None else ""
-        return f"{A}:{B}"
-
-
-class FloatMatch(AlternativeMatch):
-    def __init__(self, matchstr: str):
-        try:
-            if matchstr.count(":") != 1:
-                raise ValueError
-            first, second = [p.strip() for p in matchstr.split(":")]
-            self.start = float(first) if first else float("-inf")
-            self.end = float(second) if second else float("inf")
-        except ValueError:
-            raise SpecificationError(
-                'Float match string should be of the form "A:B" where A and B can be parsed as floats or left empty'
-            )
-
-    def check(self, val) -> bool:
-        return isinstance(val, float) and self.start <= val <= self.end
-
-    def __str__(self):
-        A = str(self.start) if self.start != float("-inf") else ""
-        B = str(self.end) if self.end != float("inf") else ""
-        return f"{A}:{B}"
-
-
-class BoolMatch(AlternativeMatch):
-    def __init__(self, matchstr: str | bool):
-        if isinstance(matchstr, bool):
-            self.val = matchstr
-            return
-        matchstr = matchstr.strip().lower()
-        if matchstr not in {"true", "false"}:
-            raise SpecificationError(
-                'Bool match string should be either "true" or "false"'
-            )
-        self.val = {"true": True, "false": False}[matchstr]
-
-    def check(self, val) -> bool:
-        return isinstance(val, bool) and val == self.val
-
-    def __str__(self):
-        return str(self.val)
+from .general import SpecificationError, is_copyfrom, type_field_mapping, type_mapping
+from .matcher import AlternativeMatch
+from .config_path import Path
+from .parser import Parser, DefaultObjectParser
 
 
 class Metadata:
@@ -394,467 +142,3 @@ class Metadata:
         pass
 
 
-class Parser:
-    NAME: str = ""
-    OUTPUT_TYPE: str = ""
-    
-    @staticmethod
-    def get_dependencies() -> list[Path]:
-        return []
-
-    def __init__(
-        self,
-        data: dict,
-        specification: dict,
-        path: Path,
-        warning_func: Callable,
-        error_func: Callable,
-    ):
-        self.data = data
-        self.specification = specification
-        self.path = path
-        self.spec_path = path.spec_path()
-        self.warning_func = warning_func
-        self.error_func = error_func
-
-        if not self.NAME:
-            raise NotImplementedError(
-                "Subclasses of Parser need to set the name of the parsing rule"
-            )
-        if not self.OUTPUT_TYPE:
-            raise NotImplementedError(
-                "Subclasses of Parser need to set the output type of the parsing rule"
-            )
-
-        required_type = self.spec_path.index(specification)["type"]
-        if required_type != self.OUTPUT_TYPE:
-            raise SpecificationError(
-                f"Parsing rule for {path} outputs {self.OUTPUT_TYPE}, but the output should be of type {required_type}"
-            )
-
-        if self.OUTPUT_TYPE in ("string", "int", "float", "bool"):
-            alternatives = Path.combine(self.spec_path, "alternatives").index(
-                self.specification
-            )
-            if alternatives is None:
-                self.alternatives = None
-            else:
-                self.alternatives = [
-                    (AlternativeMatch.get_matcher(self.OUTPUT_TYPE, key), val)
-                    for key, val in alternatives.items()
-                ]
-        else:
-            self.alternatives = None
-
-    def parse(self):
-        val = self.path.index(self.data)
-        out = self._parse(val)
-
-        flags = Path.combine(self.spec_path, "flags").index(self.specification) or []
-        if "deprecated" in flags and out is not None:
-            self.warning_func(f"deprecated property was provided ({self.path})")
-
-        if self.alternatives is not None:
-            found_match = False
-            for matcher, checks in self.alternatives:
-                if matcher.check(out):
-                    found_match = True
-                    if "warn" in checks:
-                        self.warning_func(checks["warn"])
-            if not found_match:
-                alts = ", ".join(f'"{matcher}"' for matcher in self.alternatives.keys())
-                self.error_func(
-                    f"Property {self.path} did not match any of the specified alternatives ({alts})"
-                )
-                out = None
-
-        if out is None:
-            fallback = Path.combine(self.spec_path, "default").index(self.specification)
-            if fallback is not None:
-                if isinstance(fallback, str) and fallback.startswith("copy-from:"):
-                    fallback = ("copy-from", Path.parse(fallback.split(":")[1]))
-                return fallback
-            return type_mapping[self.OUTPUT_TYPE]()
-
-        if not (
-            isinstance(out, tuple) or isinstance(out, type_mapping[self.OUTPUT_TYPE])
-        ):
-            raise SpecificationError(
-                f'Parsing rule "{self.NAME}" did not output the correct type. Output was: {out}'
-            )
-        return out
-
-    def _parse(self, val):
-        raise NotImplementedError("Subclasses of Parse need to implement _parse()")
-
-    def smallest_edit_dist(a: str, b: list[str]) -> str:
-        def edit_dist(a: str, b: str) -> int:
-            n = len(a)
-            m = len(b)
-            dp = [[0] * (m + 1) for _ in range(n + 1)]
-            for i in range(n + 1):
-                dp[i][0] = i
-            for j in range(m + 1):
-                dp[0][j] = j
-            for i in range(1, n + 1):
-                for j in range(1, m + 1):
-                    if a[i - 1] == b[j - 1]:
-                        dp[i][j] = dp[i - 1][j - 1]
-                    else:
-                        dp[i][j] = 1 + min(dp[i - 1][j], dp[i][j - 1], dp[i - 1][j - 1])
-            return dp[n][m]
-        best = b[0]
-        best_dist = edit_dist(a, best)
-        for s in b[1:]:
-            dist = edit_dist(a, s)
-            if dist < best_dist:
-                best = s
-                best_dist = dist
-        return best
-
-    @staticmethod
-    def get_parser_type(specification: dict) -> type:
-        parsing_rule = specification.get("parsing")
-        if parsing_rule is None:
-            typ = specification.get("type")
-
-            if typ is None:
-                had = "', '".join(specification.keys())
-                raise SpecificationError(
-                    f"Specification did not have a MUST HAVE field 'type', the provided fields were: ('{had}')"
-                )
-
-            if typ not in type_mapping:
-                valid = "', '".join(type_mapping.keys())
-                closest = Parser.smallest_edit_dist(typ, [*type_mapping.keys()])
-                raise SpecificationError(
-                    f"Type '{typ}' is not a valid type. Did you mean: '{closest}'? Otherwise valid types are: ('{valid}')"
-                )
-
-
-            fields = specification.keys()
-            allowed_fields = type_field_mapping.get(typ, []) + type_field_mapping["*"]
-            for field in fields:
-                if field not in allowed_fields:
-                    raise SpecificationError(
-                        f"Field '{field}' is not allowed for type '{typ}', did you mean '{Parser.smallest_edit_dist(field, allowed_fields)}'?"
-                    )
-
-            parsing_rule = f"default-{typ}-parser"
-        if parsing_rule not in parsers:
-            raise SpecificationError(f'Parser "{parsing_rule}" is not implemented')
-        return parsers[parsing_rule]
-
-
-class DefaultStringParser(Parser):
-    NAME = "default-string-parser"
-    OUTPUT_TYPE = "string"
-
-    def _parse(self, val):
-        if val is None:
-            return None
-        if not isinstance(val, str):
-            self.warning_func(
-                f"Expected value of type string but got {val}, casting to string ({self.path})"
-            )
-            val = str(val)
-        return val
-
-
-class DefaultObjectParser(Parser):
-    NAME = "default-object-parser"
-    OUTPUT_TYPE = "object"
-
-    def _parse(self, val):
-        if val is None:
-            return None
-
-        if not isinstance(val, dict):
-            self.error_func(f"Expected an object, got {val} ({self.path})")
-            return None
-
-        required = (
-            Path.combine(self.spec_path, "required").index(self.specification) or []
-        )
-        for req in required:
-            req_path = Path.combine(self.path, req)
-            if req_path.index(self.data) is None:
-                self.error_func(f"Missing required property: {req_path}")
-
-        remove = []
-        known_props = Path.combine(self.spec_path, "properties").index(
-            self.specification
-        )
-        for prop in val.keys():
-            if prop not in known_props:
-                self.warning_func(f"Unknown property: {Path.combine(self.path, prop)}")
-                remove.append(prop)
-        for r in remove:
-            del val[r]
-
-        return val
-
-
-class DefaultListParser(Parser):
-    NAME = "default-list-parser"
-    OUTPUT_TYPE = "list"
-
-    def _parse(self, val):
-        if val is None:
-            return None
-
-        if not isinstance(val, list):
-            self.error_func(f"Expected a list, got {val} ({self.path})")
-            return None
-
-        return val
-
-
-class DefaultIntParser(Parser):
-    NAME = "default-int-parser"
-    OUTPUT_TYPE = "int"
-
-    def _parse(self, val):
-        if val is None:
-            return None
-
-        if not isinstance(val, int):
-            try:
-                cast = int(val)
-                self.warning_func(
-                    f"Expected type int, got {val}. Casting to {cast} ({self.path})"
-                )
-                val = cast
-            except ValueError:
-                self.error_func(f"Expected a int, got {val} ({self.path})")
-                return None
-
-        return val
-
-
-class DefaultFloatParser(Parser):
-    NAME = "default-float-parser"
-    OUTPUT_TYPE = "float"
-
-    def _parse(self, val):
-        if val is None:
-            return None
-
-        if not isinstance(val, float):
-            try:
-                cast = float(val)
-                self.warning_func(
-                    f"Expected type float, got {val}. Casting to {cast} ({self.path})"
-                )
-                val = cast
-            except ValueError:
-                self.error_func(f"Expected a float, got {val} ({self.path})")
-                return None
-
-        return val
-
-
-class DefaultBoolParser(Parser):
-    NAME = "default-bool-parser"
-    OUTPUT_TYPE = "bool"
-
-    def _parse(self, val):
-        if val is None:
-            return None
-
-        if isinstance(val, str):
-            if val.lower() in ("true", "false"):
-                interpretation = val.lower() == "true"
-                self.warning_func(
-                    f'Expected type bool but got a string "{val}" which will be interpreted as {interpretation} ({self.path})'
-                )
-                val = interpretation
-            else:
-                self.error_func(f'Expected type bool but got "{val}" ({self.path})')
-                return None
-
-        if not isinstance(val, bool):
-            self.error_func(f"Expected type bool, got {val} ({self.path})")
-            return None
-
-        return val
-
-
-parsers = {
-    p.NAME: p
-    for p in [
-        DefaultObjectParser,
-        DefaultListParser,
-        DefaultStringParser,
-        DefaultIntParser,
-        DefaultFloatParser,
-        DefaultBoolParser,
-    ]
-}
-
-
-class BaseValidator:
-    def __init__(self, layout: dict, metadata: Metadata, path: str = ""):
-        self.layout = layout
-        self.metadata = metadata
-        self.path = path
-
-    def verify(self, value):
-        """
-        Verifies the value:
-          - Applies defaults
-          - Converts types
-          - Logs warnings/errors if needed
-        """
-        raise NotImplementedError("Subclasses must implement verify")
-
-    def check(self, value):
-        """
-        Performs extra-checks (like forbid/require logic)
-        get_path_func can be used to fetch other values by path.
-        """
-        raise NotImplementedError("Subclasses must implement check")
-
-
-class StringValidator(BaseValidator):
-    def __init__(self, layout: dict, metadata: Metadata, path: str = ""):
-        super().__init__(layout, metadata, path)
-        alternatives = self.layout.get("alternatives")
-        if alternatives:
-            self.patterns = {alt: re.compile("^" + alt + "$") for alt in alternatives}
-        else:
-            self.patterns = None
-
-    def verify(self, value):
-        if value is None:
-            value = self.layout.get("default", "")
-        if not isinstance(value, str):
-            self.metadata.warning_func(
-                f"Property {self.path} was expected to be of type string"
-            )
-            value = str(value)
-        if self.patterns:
-            if not any(pattern.match(value) for pattern in self.patterns.values()):
-                self.metadata.error_func(
-                    f"Property {self.path} is {value} but must match one of {list(self.patterns.keys())}"
-                )
-                value = self.layout.get("default", "")
-        return value
-
-    def check(self, value):
-        if not self.patterns:
-            return
-        match = next(
-            (key for key, pattern in self.patterns.items() if pattern.match(value)),
-            None,
-        )
-        checks = self.layout["alternatives"][match]
-        for forbidden in checks.get("forbid", []):
-            other_path, expected = forbidden.split(":")
-            if self.metadata[other_path] == expected:
-                self.metadata.error_func(
-                    f"Property {self.path} has value {value} which forbids property {other_path} to have value {expected}"
-                )
-        for required in checks.get("required", []):
-            if not self.metadata[
-                required
-            ]:  # TODO: This is not a good way to handle this check I think
-                self.metadata.error_func(
-                    f"Property {self.path} has value {value} which requires property {required}"
-                )
-        if "warn" in checks:
-            self.metadata.warning_func(checks["warn"])
-
-
-class ObjectValidator(BaseValidator):
-    def verify(self, value):
-        if value is None:
-            return self.layout.get("default", {})
-        if self.layout.get("parsing") == "legacy-validation":
-            if not isinstance(value, str):
-                self.metadata.error_func(
-                    f"Property {self.path} was expected to be a string"
-                )
-                return {}
-            elements = value.split()
-            value = {
-                "type": elements[0],
-                "interactive": "interactive" in elements[1:],
-                "score": "score" in elements[1:],
-            }
-        if not isinstance(value, dict):
-            self.metadata.error_func(
-                f"property {self.path} was expected to be a dictionary"
-            )
-            return {}
-        for prop in self.layout.get("required", []):
-            if prop not in value:
-                self.metadata.error_func(
-                    f"Missing required property: {self.path}/{prop}"
-                )
-        for prop in value.keys():
-            if prop not in self.layout["properties"]:
-                self.metadata.warning_func(f"Unknown property: {self.path}/{prop}")
-        return value
-
-    def check(self, value):
-        pass
-
-
-class ListValidator(BaseValidator):
-    def verify(self, value):
-        if value is None:
-            return self.layout.get("default", [])
-        if self.layout.get("parsing") == "space-separated-strings":
-            if not isinstance(value, str):
-                self.metadata.error_func(
-                    f"Property {self.path} was expected to be a string"
-                )
-                return []
-            value = value.split()
-        if not isinstance(value, list):
-            self.metadata.error_func(f"property {self.path} was expected to be a list")
-            return []
-        return value
-
-    def check(self, value):
-        pass
-
-
-class FloatValidator(BaseValidator):
-    def verify(self, value):
-        if value is None:
-            return self.layout.get("default", 0.0)
-        if not isinstance(value, float):
-            try:
-                value = float(value)
-            except Exception:
-                self.metadata.error_func(
-                    f"Property {self.path} was expected to be a float"
-                )
-                value = self.layout.get("default", 0.0)
-        return value
-
-    def check(self, value):
-        pass
-
-
-class IntValidator(BaseValidator):
-    def verify(self, value):
-        if value is None:
-            return self.layout.get("default", 0)
-        if not isinstance(value, int):
-            try:
-                value = int(value)
-                self.metadata.warning_func(
-                    f"Property {self.path} should be of type integer, interpreting as {value}"
-                )
-            except Exception:
-                self.metadata.error_func(
-                    f"Property {self.path} was expected to be an integer"
-                )
-                value = self.layout.get("default", 0)
-        return value
-
-    def check(self, value):
-        pass
